# optimize ECAL uncalibRecHit multifit (follow up to #18600)