--- a/RecoLocalCalo/EcalRecProducers/interface/EcalUncalibRecHitWorkerBaseClass.h
+++ b/RecoLocalCalo/EcalRecProducers/interface/EcalUncalibRecHitWorkerBaseClass.h
@@ -23,6 +23,14 @@ class EcalUncalibRecHitWorkerBaseClass {
   virtual void set(const edm::EventSetup& es) = 0;
   virtual void set(const edm::Event& evt) {}
   virtual bool run(const edm::Event& evt, const EcalDigiCollection::const_iterator & digi, EcalUncalibratedRecHitCollection & result) = 0;
+
+  virtual void run(const edm::Event& evt, const EcalDigiCollection & digis, EcalUncalibratedRecHitCollection & result)
+  {
+    result.reserve(result.size() + digis.size());
+    for (auto it = digis.begin(); it != digis.end(); ++it)
+      run(evt, it, result);
+  }
+
   virtual edm::ParameterSetDescription getAlgoDescription() = 0;
 };