… RecoLocalCalo/EcalRecAlgos/BuildFile.xml
… RecoLocalCalo/EcalRecProducers/plugins/EcalUncalibRecHitWorkerMultiFit.h
… RecoLocalCalo/EcalRecProducers/plugins/EcalUncalibRecHitWorkerRatio.cc
… RecoLocalCalo/EcalRecProducers/plugins/EcalUncalibRecHitWorkerRatio.h
… RecoLocalCalo/EcalRecProducers/plugins/EcalUncalibRecHitWorkerWeights.h