From #18600 

```
loop over DIGIs inside EcalUncalibRecHitWorker*::run()
compile EcalRecProducers and EcalRecAlgos with -O3 <== SK: this became just a default flag during review of #18600

from @felicepantaleo :

    faster check for a positive vector
    increase threshold with a smoother granularity

from @VinInn :

    use lazyProduct() and noalias() in NNLS

```

The original was not signed off due to remaining duplication of rechit creation implementation in EcalUncalibRecHitWorkerMultiFit.


As a follow up to the original submission, the following changes were made https://github.com/cms-sw/cmssw/compare/158d3b699bee13299384f2a3df528013a00cc3d9...670bfda05db3bdf761548c4c0d9f20cc2b546480:
- treat the proposed ::run method using all digis as the new main default for running workers as seen by EcalUncalibRecHitProducer
    - this becomes the only ::run method for EcalUncalibRecHitWorkerBaseClass and its direct descendants. This automatically allows to remove the single-digi ::run method from EcalUncalibRecHitWorkerMultiFit.
    -  introduce a compatibility base EcalUncalibRecHitWorkerRunOneDigiBase  which calls this for old methods running on their individual digi ::run
- cleanup of leftover commented out lines

Tested with regular short matrix (includes non-multiFit ecal reco)
and also for timing in JetHT run 282730 LS 160 (PU~45). There are no differences in reconstructed objects.
For the timing of ecalMultiFitUncalibRecHit
-  there are no significant difference in CPU time  compared to #18600  (I also checked re-HLT of the same input and time of hltEcalUncalibRecHit to be sure)
- there is about 12% improvement compared to the baseline CMSSW_9_2_X_2017-05-17-1100; in agreement with https://patatrack.web.cern.ch/content/ecal-multifit-hlt-optimization


@argiro @fwyzard 
please check if this update is acceptable to you