@cmsbuild 2017-05-18
The tests are being triggered in jenkins.