@cmsbuild 2017-05-19
+1
Tested at: 670bfda05db3bdf761548c4c0d9f20cc2b546480
https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-18840/19977/summary.html