@cmsbuild 2017-05-19
Comparison is ready
https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-18840/19977/summary.html

Comparison Summary:
 - You potentially added 141 lines to the logs
 - Reco comparison results: 3323 differences found in the comparisons
 - DQMHistoTests: Total files compared: 24
 - DQMHistoTests: Total histograms compared: 1833935
 - DQMHistoTests: Total failures: 33196
 - DQMHistoTests: Total nulls: 1
 - DQMHistoTests: Total successes: 1800558
 - DQMHistoTests: Total skipped: 180
 - DQMHistoTests: Total Missing objects: 0
 - Checked 98 log files, 14 edm output root files, 24 DQM output files