@fwyzard 2017-05-19
To me this approach looks rather baroque, but at least it is better than both proposals made in #18600 . 
I can confirm the same performance gain as in #18600 :

| code base | average time | gain |
|-------------|-----------------|-----------|
| CMSSW_9_2_0 |  34.3 ± 0.1 ms   | reference |
| #18600      |  29.16 ± 0.04 ms | -15.2%    | 
| #18840      |  29.20 ± 0.05 ms | -15.1%    |

So, OK from my side.