@slava77 2017-05-19
On 5/19/17 2:32 AM, Andrea Bocci wrote:
> To me this approach looks rather baroque, but at least it is better than
> both proposals made in #18600 <https://github.com/cms-sw/cmssw/pull/18600> .

Yes, it is baroque,
but it seemed more straightforward than the duality in the interface of
the lowest level base class
with a dead-ended ::run(singleDigi for multifit.
A less baroque solution requires migration of the rest of the derived
classes.


> I can confirm the same performance gain as in #18600
> <https://github.com/cms-sw/cmssw/pull/18600> :
> 
> code base 	average time 	gain
> CMSSW_9_2_0 	34.3 ± 0.1 ms 	reference
> #18600 <https://github.com/cms-sw/cmssw/pull/18600> 	29.16 ± 0.04 ms
> -15.2%
> #18840 <https://github.com/cms-sw/cmssw/pull/18840> 	29.20 ± 0.05 ms
> -15.1%
> 

Thank you for confirming.


> So, OK from my side.
> 
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/cms-sw/cmssw/pull/18840#issuecomment-302655999>, or
> mute the thread
> <https://github.com/notifications/unsubscribe-auth/AEdcbsO9gh_mhDT-Ypyso-ZG77qitNs4ks5r7WGfgaJpZM4Nf3NC>.
>