@slava77 2017-05-19
+1

for https://github.com/cms-sw/cmssw/pull/18840	670bfda
- jenkins tests pass and comparisons with baseline show no differences as expected with pre-signoff tests